rename result2 to result
